fix(Camera): use the swapchain aspect ratio for the projection

+++ include/Swapchain.hpp
@@ -19,7 +19,10 @@ public:
     uint32_t nbOfImage() const;
 
     constexpr const VkSwapchainKHR &getSwapchain() const { return swapChain; }
-    constexpr float getAspectRatio() const { return swapChainExtent.width / swapChainExtent.height; }
+    constexpr float getAspectRatio() const
+    {
+        return static_cast<float>(swapChainExtent.width) / static_cast<float>(swapChainExtent.height);
+    }
     constexpr VkImage &getSwapchainImage(const unsigned index) { return swapChainImages.at(index); }
     constexpr VkImageView &getSwapchainImageView(const unsigned index) { return swapChainImageViews.at(index); }
     constexpr const VkFormat &getSwapchainFormat() const { return swapChainImageFormat; }
